✨ update Customer and Product models to use Integer for IDs and add repositories for data management

+++ src/main/java/com/example/demo/model/Customer.java
@@ -1,22 +1,22 @@
 package com.example.demo.model;
 
 public class Customer{
-    private Long id;
+    private Integer id;
     private String firstName;
     private String lastName;
     private String email;
 
-    public Customer(Long id, String firstName, String lastName, String email) {
+    public Customer(Integer id, String firstName, String lastName, String email) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
     }
-    
-    public Long getId() {
+
+    public Integer getId() {
         return id;
     }
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
     public String getFirstName() {
